Enum squid::download::Error

pub enum Error {
    InvalidOption(String),
    Http(Error),
    Tls(Error),
    MissingResponseHeader(String),
    UnexpectedResponseLength {
        expected: u64,
        actual: u64,
    },
    EmptyTarget(String),
    Io(Error),
    Internal,
}

Enumerates possible download errors.

Variants
InvalidOption(String)

An option is not properly configured.

Http(Error)

An HTTP-related error has occurred.

Tls(Error)

A TLS-related error has occurred.

MissingResponseHeader(String)

The response from the server was missing a header.

UnexpectedResponseLength

The response from the server has different length than expected.

Fields of UnexpectedResponseLength

expected: u64

The expected number of bytes to receive.

actual: u64

The actual number of bytes received.

EmptyTarget(String)

The target's content length is zero bytes.

Io(Error)

A filesystem IO-related error has occurred.

Internal

An internal error has occurred.